server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Werkzeug==2.3.8
SQLAlchemy==2.0.25
openpyxl==3.1.2
python-docx==1.1.0
APScheduler==3.10.4
